pyproject.toml
[project]
name = "gpm-django-be"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.2.8,<6.0.0)",
    "typeguard (>=4.4.4,<5.0.0)",
    "django-cors-headers (>=4.9.0,<5.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "dj-rest-auth (>=7.0.1,<8.0.0)",
    "django-allauth (>=65.13.1,<66.0.0)",
    "drf-yasg (>=1.21.11,<2.0.0)",
    "requests (>=2.32.5,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest-django (>=4.11.1,<5.0.0)",
    "mixer (>=7.2.2,<8.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)",
    "pytest (>=9.0.1,<10.0.0)"
]
